Fix getJoinedValues for global and viewpor histogram by using different methods
